feat: use mappers to determine http protocol and port numbers instead
